requirements.txt
alembic==1.13.1
blinker==1.7.0
cachelib==0.12.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
cryptography==42.0.5
Flask==3.0.2
Flask-Migrate==4.0.5
Flask-MySQLdb==2.0.0
Flask-Session==0.4.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
gunicorn==20.1.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.2
MarkupSafe==2.1.5
mysql-connector-python==8.3.0
mysqlclient==2.2.4
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==0.19.2
requests==2.27.1
SQLAlchemy==2.0.27
typing_extensions==4.9.0
urllib3==1.26.18
Werkzeug==3.0.1
youtube-dl==2021.12.17
